impl: `hdp-preprocessor` with datalake compile
