+++ crates/core/src/lib.rs
@@ -1,10 +1,6 @@
-#![deny(unused_crate_dependencies)]
-
 pub mod cairo_runner;
-pub mod compiler;
 pub mod config;
 pub mod constant;
 pub mod conversion;
 pub mod module_registry;
-pub mod pre_processor;
 pub mod processor;
