+++ crates/pre-processor/Cargo.toml
@@ -1,32 +1,22 @@
 [package]
-name = "hdp-core"
-description = "HDP core implementation"
+name = "hdp-preprocessor"
+version.workspace = true
 edition.workspace = true
 license-file.workspace = true
+authors.workspace = true
 repository.workspace = true
-version.workspace = true
+homepage.workspace = true
 exclude.workspace = true
 keywords.workspace = true
 categories.workspace = true
-authors.workspace = true
-readme = "README.md"
 
 [dependencies]
-hdp-preprocessor = { workspace = true }
 hdp-provider = { workspace = true }
 hdp-primitives = { workspace = true }
 alloy-merkle-tree = { workspace = true }
 alloy = { workspace = true }
 anyhow = { workspace = true }
 cairo-lang-starknet-classes.workspace = true
-starknet-crypto.workspace = true
-starknet.workspace = true
 tokio = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-serde_with = { workspace = true }
 tracing = { workspace = true }
-lazy_static = { workspace = true }
-regex = { workspace = true }
-tempfile = { workspace = true }
-futures = { workspace = true }
+thiserror = { workspace = true }
