+++ crates/pre-processor/src/compile/datalake/mod.rs
@@ -1,89 +1,25 @@
-use alloy::primitives::{B256, U256};
-use anyhow::{bail, Result};
+use std::collections::{HashMap, HashSet};
+
+use crate::compile::datalake::fetchable::Fetchable;
 use hdp_primitives::{
     processed_types::{
-        account::ProcessedAccount, header::ProcessedHeader, mmr::MMRMeta,
-        receipt::ProcessedReceipt, storage::ProcessedStorage, transaction::ProcessedTransaction,
+        account::ProcessedAccount, header::ProcessedHeader, receipt::ProcessedReceipt,
+        storage::ProcessedStorage, transaction::ProcessedTransaction,
     },
     solidity_types::traits::DatalakeComputeCodecs,
     task::datalake::{envelope::DatalakeEnvelope, DatalakeCompute},
 };
 use hdp_provider::evm::provider::{EvmProvider, EvmProviderConfig};
-use std::{
-    collections::{HashMap, HashSet},
-    sync::Arc,
-};
-use tokio::sync::RwLock;
 
-use self::{block_sampled::compile_block_sampled_datalake, transactions::compile_tx_datalake};
+use super::{Compilable, CompilationResults, CompileError};
 
-pub mod block_sampled;
-pub mod transactions;
+pub mod fetchable;
 
-pub struct DatalakeComputeCompilationResults {
-    /// flag to check if the aggregation function is pre-processable
-    pub pre_processable: bool,
-    /// task_commitment -> value
-    pub commit_results_maps: HashMap<B256, U256>,
-    /// Headers related to the datalake
-    pub headers: HashSet<ProcessedHeader>,
-    /// Accounts related to the datalake
-    pub accounts: HashSet<ProcessedAccount>,
-    /// Storages related to the datalake
-    pub storages: HashSet<ProcessedStorage>,
-    /// Transactions related to the datalake
-    pub transactions: HashSet<ProcessedTransaction>,
-    /// Transaction receipts related to the datalake
-    pub transaction_receipts: HashSet<ProcessedReceipt>,
-    /// MMR meta data related to the headers
-    pub mmr_meta: MMRMeta,
-}
-
-impl DatalakeComputeCompilationResults {
-    #[allow(clippy::too_many_arguments)]
-    pub fn new(
-        pre_processable: bool,
-        commit_results_maps: HashMap<B256, U256>,
-        headers: HashSet<ProcessedHeader>,
-        accounts: HashSet<ProcessedAccount>,
-        storages: HashSet<ProcessedStorage>,
-        transactions: HashSet<ProcessedTransaction>,
-        transaction_receipts: HashSet<ProcessedReceipt>,
-        mmr_meta: MMRMeta,
-    ) -> Self {
-        Self {
-            pre_processable,
-            commit_results_maps,
-            headers,
-            accounts,
-            storages,
-            transactions,
-            transaction_receipts,
-            mmr_meta,
-        }
-    }
-}
-
-pub struct DatalakeCompiler {
-    provider: Arc<RwLock<EvmProvider>>,
-}
-
-impl DatalakeCompiler {
-    /// initialize DatalakeCompiler with commitment and datalake
-    pub fn new_from_config(config: EvmProviderConfig) -> Self {
-        let provider = EvmProvider::new(config);
-        Self {
-            provider: Arc::new(provider.into()),
-        }
-    }
-
-    /// Compile the datalake meaning, fetching relevant headers, accounts, storages, and mmr_meta data.
-    ///
-    /// Plus, it will combine target datalake's datapoints in compiled_results.
-    pub async fn compile(
+impl Compilable for Vec<DatalakeCompute> {
+    async fn compile(
         &self,
-        datalake_computes: &[DatalakeCompute],
-    ) -> Result<DatalakeComputeCompilationResults> {
+        provider_config: &EvmProviderConfig,
+    ) -> Result<CompilationResults, CompileError> {
         let mut commit_results_maps = HashMap::new();
         let mut headers: HashSet<ProcessedHeader> = HashSet::new();
         let mut accounts: HashSet<ProcessedAccount> = HashSet::new();
@@ -93,19 +29,19 @@ impl DatalakeCompiler {
         let mut mmr = None;
         let mut pre_processable = true;
 
-        for datalake_compute in datalake_computes {
+        for datalake_compute in self {
             let task_commitment = datalake_compute.commit();
             let aggregation_fn = &datalake_compute.compute.aggregate_fn_id;
             let fn_context = datalake_compute.compute.aggregate_fn_ctx.clone();
+            let provider = EvmProvider::new(provider_config.clone());
             match datalake_compute.datalake {
                 DatalakeEnvelope::BlockSampled(ref datalake) => {
-                    let compiled_block_sampled =
-                        compile_block_sampled_datalake(datalake.clone(), &self.provider).await?;
+                    let compiled_block_sampled = datalake.fetch(provider).await?;
                     headers.extend(compiled_block_sampled.headers);
                     accounts.extend(compiled_block_sampled.accounts);
                     storages.extend(compiled_block_sampled.storages);
                     if mmr.is_some() && mmr.unwrap() != compiled_block_sampled.mmr_meta {
-                        bail!("MMR meta data is not consistent");
+                        return Err(CompileError::InvalidMMR);
                     } else {
                         mmr = Some(compiled_block_sampled.mmr_meta);
                     }
@@ -120,14 +56,13 @@ impl DatalakeCompiler {
                     }
                 }
                 DatalakeEnvelope::Transactions(ref datalake) => {
-                    let compiled_tx_datalake =
-                        compile_tx_datalake(datalake.clone(), &self.provider).await?;
+                    let compiled_tx_datalake = datalake.fetch(provider).await?;
                     headers.extend(compiled_tx_datalake.headers);
                     transactions.extend(compiled_tx_datalake.transactions);
                     transaction_receipts.extend(compiled_tx_datalake.transaction_receipts);
 
                     if mmr.is_some() && mmr.unwrap() != compiled_tx_datalake.mmr_meta {
-                        bail!("MMR meta data is not consistent");
+                        return Err(CompileError::InvalidMMR);
                     } else {
                         mmr = Some(compiled_tx_datalake.mmr_meta);
                     }
@@ -144,7 +79,7 @@ impl DatalakeCompiler {
             };
         }
 
-        Ok(DatalakeComputeCompilationResults::new(
+        Ok(CompilationResults::new(
             pre_processable,
             commit_results_maps,
             headers,
